+++ src/main/java/bgu/spl/net/OP4LogoutMessage.java
@@ -14,12 +14,12 @@ public class OP4LogoutMessage implements OPMessage {
         this.loggedInUser = s;
         Database database = Database.getInstance();
         if (loggedInUser == null) {
-            return new OP13ErrMessage(13, 4);
+            return new OP13ErrMessage(13, (short) 4);
         }
         synchronized (database.getUsersInfo().get(loggedInUser).isLoggedIn()) {
             database.getUsersInfo().get(loggedInUser).logOut();
         }
-        return new OP12AckMessage(12, 4, "");
+        return new OP12AckMessage(12, (short) 4, "");
     }
 
     @Override
