+++ src/main/java/bgu/spl/net/OP7PrintCourseStatMessage.java
@@ -17,13 +17,13 @@ public class OP7PrintCourseStatMessage implements OPMessage {
         this.loggedInUser = s;
         Database database = Database.getInstance();
         if (loggedInUser == null) {
-            return new OP13ErrMessage(13, 7);
+            return new OP13ErrMessage(13, (short) 7);
         }
         if (!(database.getUsersInfo().get(loggedInUser).isAdmin())){
-            return new OP13ErrMessage(13, 7);
+            return new OP13ErrMessage(13, (short) 7);
         }
         if (!(database.getCoursesInfo().containsKey(courseNum))){
-            return new OP13ErrMessage(13, 7);
+            return new OP13ErrMessage(13, (short) 7);
         }
         synchronized (database.getCoursesInfo().get(courseNum).getCurrStudents()) {
             Course c = database.getCoursesInfo().get(courseNum);
@@ -31,7 +31,7 @@ public class OP7PrintCourseStatMessage implements OPMessage {
             String stat = "Course:(" + courseNum + ")" + c.getCourseName() + "\n"
                     + "Seats Available:" + c.getCurrStudents() + "/"
                     + c.getMaxStudents() + "\n" + "Student Registered:" + studReg;
-            return new OP12AckMessage(12, 7, stat);
+            return new OP12AckMessage(12, (short) 7, stat);
         }
     }
