+++ src/main/java/bgu/spl/net/OP8PrintStudStatMessage.java
@@ -17,18 +17,18 @@ public class OP8PrintStudStatMessage implements OPMessage {
         this.loggedInUser = s;
         Database database = Database.getInstance();
         if (loggedInUser == null) {
-            return new OP13ErrMessage(13, 8);
+            return new OP13ErrMessage(13, (short) 8);
         }
         if (!(database.getUsersInfo().get(loggedInUser).isAdmin())){
-            return new OP13ErrMessage(13, 8);
+            return new OP13ErrMessage(13, (short) 8);
         }
         if (!(database.getUsersInfo().containsKey(studUsername))){
-            return new OP13ErrMessage(13, 8);
+            return new OP13ErrMessage(13, (short) 8);
         }
         User u = database.getUsersInfo().get(studUsername);
         String regCourses = Arrays.toString(u.getRegisteredCourses().toArray());
         String stat = "Student:"+studUsername +"\n" +"Courses:"+regCourses;
-        return new OP12AckMessage(12,8,stat);
+        return new OP12AckMessage(12, (short) 8,stat);
     }
 
     @Override
